# Improve performance of mutable proto structs