--- a/src/ProtoStruct.jl
+++ b/src/ProtoStruct.jl
@@ -72,7 +72,7 @@ macro proto(expr)
     const_field_names = [f for (f, fi) in zip(field_names, field_info) if fi[3] == true]
 
     if ismutable
-        field_types = :(Tuple{$((:(Base.RefValue{<:$x}) for x in getindex.(field_info, 2))...)})
+        field_types = :(Tuple{$((:(Base.RefValue{$x}) for x in getindex.(field_info, 2))...)})
         fields_with_ref = (:($x=Ref($x)) for x in field_names)
     else
         field_types = :(Tuple{$(getindex.(field_info, 2)...)})