@Tortar 2023-10-18
Now it works for all types, verified that the performance is still on pair with parametric types